refactor: page, layout, globalError  TS 마이그레이션
