+++ .gitignore
@@ -44,3 +44,4 @@ next-env.d.ts
 
 # etc
 react-project
+jsconfig.json
