+++ src/app/community/[articleId]/page.tsx
