+++ src/app/products/[productId]/page.tsx
